pyproject.toml
[project]
name = "ddgem"
version = "0.1.0"
description = "A library to get fact-checked answers from Gemini LLM using DDG search."
authors = [
    {name = "Slava Beatum",email = "elrifpro@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
packages = [{include = "ddgem", from = "src"}]


requires-python = ">= 3.10"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-genai (>=1.27.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "ddgs (>=9.4.3,<10.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.1"
pytest = "^8.4.1"